+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.5.0] - 2022-06-22
 
 ## Changed
+ - Rune() API change: unique_id comes first, and constructor generates a valid Rune from a master authcode() as expected.
  - stricter checks on use of unique_id field to avoid caller having to.
 
 ### Added